requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.0.13
idna==2.8
nudepy==0.4
Pillow==6.2.0
pkg-resources==0.0.0
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6
